Issue9/embedding-service/requirements.txt
confluent_kafka
weaviate-client
torch
timm
pillow
